Refactor building of rejected error message
The message is now built by the `ComponentState` itself, which knows
_why_ it was rejected.

Issue #8428
